+++ src/main/java/br/com/zalf/prolog/webservice/integracao/avacorpavilan/AvaCorpAvilanConstants.java
@@ -1,0 +1,19 @@
+package br.com.zalf.prolog.webservice.integracao.avacorpavilan;
+
+/**
+ * Created on 9/28/17
+ *
+ * @author Luiz Felipe (https://github.com/luizfp)
+ */
+public class AvaCorpAvilanConstants {
+
+    private AvaCorpAvilanConstants() {
+
+    }
+
+    public static final String BASE_URL = "http://189.11.175.146/IntegracaoPrologTestes/";
+//    public static final String BASE_URL = "http://189.11.175.146/IntegracaoProlog/";
+
+    public static final String NAMESPACE = "http://www.avacorp.com.br/integracaoprologtestes";
+//    public static final String NAMESPACE = "http://www.avacorp.com.br/integracaoprolog";
+}