+++ src/main/java/br/com/zalf/prolog/webservice/integracao/avacorpavilan/AvaCorpAvilanUtils.java
@@ -4,6 +4,7 @@ import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
 import com.sun.istack.internal.NotNull;
 
+import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
@@ -17,6 +18,9 @@ import java.util.Locale;
 @VisibleForTesting
 public class AvaCorpAvilanUtils {
 
+    private static final SimpleDateFormat AVILAN_DATE_FORMAT = new SimpleDateFormat("yyyy-MM-dd");
+    private static final SimpleDateFormat AVILAN_DATE_TIME_FORMAT = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
+
     private AvaCorpAvilanUtils() {
         throw new IllegalStateException(AvaCorpAvilanUtils.class.getSimpleName() + " cannot be instantiated!");
     }
@@ -27,8 +31,26 @@ public class AvaCorpAvilanUtils {
      * @param date um {@link Date}.
      * @return uma {@link String} represetando a data.
      */
+    @NotNull
     public static String createDatePattern(@NotNull final Date date) {
-        return new SimpleDateFormat("yyyy-MM-dd").format(date);
+        return AVILAN_DATE_FORMAT.format(date);
+    }
+
+    /**
+     * Converte uma representação em texto no padrão utilizado no web service de integração da Avilan para um
+     * {@link Date date}.
+     *
+     * @param dateString {@link String} representação da data.
+     * @return um {@link Date}.
+     */
+    @NotNull
+    public static Date createDatePattern(@NotNull final String dateString) throws ParseException {
+        return AVILAN_DATE_FORMAT.parse(dateString);
+    }
+
+    @NotNull
+    public static Date createDateTimePattern(@NotNull final String dateTimeString) throws ParseException {
+        return AVILAN_DATE_TIME_FORMAT.parse(dateTimeString);
     }
 
     public static int calculateDaysBetweenDateAndNow(@NotNull final String stringDate) {