+++ src/main/java/br/com/zalf/prolog/webservice/integracao/avacorpavilan/checklist/ObjectFactory.java
@@ -16,7 +16,7 @@ import javax.xml.bind.annotation.XmlRegistry;
  * type definitions, element declarations and model 
  * groups.  Factory methods for each of these are 
  * provided in this class.
- * 
+ *
  */
 @XmlRegistry
 public class ObjectFactory {
@@ -24,14 +24,14 @@ public class ObjectFactory {
 
     /**
      * Create a new ObjectFactory that can be used to create new instances of schema derived classes for package: br.com.avacorp.integracaoprolog
-     * 
+     *
      */
     public ObjectFactory() {
     }
 
     /**
      * Create an instance of {@link IntegrarArquivo }
-     * 
+     *
      */
     public IntegrarArquivo createIntegrarArquivo() {
         return new IntegrarArquivo();
@@ -39,7 +39,7 @@ public class ObjectFactory {
 
     /**
      * Create an instance of {@link EnviarChecklist }
-     * 
+     *
      */
     public EnviarChecklist createEnviarChecklist() {
         return new EnviarChecklist();
@@ -47,7 +47,7 @@ public class ObjectFactory {
 
     /**
      * Create an instance of {@link RespostasAvaliacao }
-     * 
+     *
      */
     public RespostasAvaliacao createRespostasAvaliacao() {
         return new RespostasAvaliacao();
@@ -55,7 +55,7 @@ public class ObjectFactory {
 
     /**
      * Create an instance of {@link BuscarPerguntasAlternativasQuestionario }
-     * 
+     *
      */
     public BuscarPerguntasAlternativasQuestionario createBuscarPerguntasAlternativasQuestionario() {
         return new BuscarPerguntasAlternativasQuestionario();
@@ -63,7 +63,7 @@ public class ObjectFactory {
 
     /**
      * Create an instance of {@link AdicionarChecklist }
-     * 
+     *
      */
     public AdicionarChecklist createAdicionarChecklist() {
         return new AdicionarChecklist();
@@ -71,7 +71,7 @@ public class ObjectFactory {
 
     /**
      * Create an instance of {@link EnviarChecklistResponse }
-     * 
+     *
      */
     public EnviarChecklistResponse createEnviarChecklistResponse() {
         return new EnviarChecklistResponse();
@@ -79,7 +79,7 @@ public class ObjectFactory {
 
     /**
      * Create an instance of {@link EnviaRespostaAvaliacao }
-     * 
+     *
      */
     public EnviaRespostaAvaliacao createEnviaRespostaAvaliacao() {
         return new EnviaRespostaAvaliacao();
@@ -87,7 +87,7 @@ public class ObjectFactory {
 
     /**
      * Create an instance of {@link BuscarQuestionariosColaborador }
-     * 
+     *
      */
     public BuscarQuestionariosColaborador createBuscarQuestionariosColaborador() {
         return new BuscarQuestionariosColaborador();
@@ -95,7 +95,7 @@ public class ObjectFactory {
 
     /**
      * Create an instance of {@link BuscarPerguntasAlternativasQuestionarioResponse }
-     * 
+     *
      */
     public BuscarPerguntasAlternativasQuestionarioResponse createBuscarPerguntasAlternativasQuestionarioResponse() {
         return new BuscarPerguntasAlternativasQuestionarioResponse();
@@ -103,7 +103,7 @@ public class ObjectFactory {
 
     /**
      * Create an instance of {@link PerguntasAlternativasQuestionario }
-     * 
+     *
      */
     public PerguntasAlternativasQuestionario createPerguntasAlternativasQuestionario() {
         return new PerguntasAlternativasQuestionario();
@@ -111,7 +111,7 @@ public class ObjectFactory {
 
     /**
      * Create an instance of {@link BuscarQuestionariosColaboradorResponse }
-     * 
+     *
      */
     public BuscarQuestionariosColaboradorResponse createBuscarQuestionariosColaboradorResponse() {
         return new BuscarQuestionariosColaboradorResponse();
@@ -119,7 +119,7 @@ public class ObjectFactory {
 
     /**
      * Create an instance of {@link BuscaQuestionarioColaborador }
-     * 
+     *
      */
     public BuscaQuestionarioColaborador createBuscaQuestionarioColaborador() {
         return new BuscaQuestionarioColaborador();
@@ -127,7 +127,7 @@ public class ObjectFactory {
 
     /**
      * Create an instance of {@link IntegrarArquivoResponse }
-     * 
+     *
      */
     public IntegrarArquivoResponse createIntegrarArquivoResponse() {
         return new IntegrarArquivoResponse();
@@ -135,15 +135,39 @@ public class ObjectFactory {
 
     /**
      * Create an instance of {@link RetornoPadrao }
-     * 
+     *
      */
     public RetornoPadrao createRetornoPadrao() {
         return new RetornoPadrao();
     }
 
     /**
+     * Create an instance of {@link FarolChecklistResponse }
+     *
+     */
+    public FarolChecklistResponse createFarolChecklistResponse() {
+        return new FarolChecklistResponse();
+    }
+
+    /**
+     * Create an instance of {@link FarolChecklist2 }
+     *
+     */
+    public FarolChecklist2 createFarolChecklist2() {
+        return new FarolChecklist2();
+    }
+
+    /**
+     * Create an instance of {@link VeiculoChecklist }
+     *
+     */
+    public VeiculoChecklist createVeiculoChecklist() {
+        return new VeiculoChecklist();
+    }
+
+    /**
      * Create an instance of {@link QuestionarioVeiculos }
-     * 
+     *
      */
     public QuestionarioVeiculos createQuestionarioVeiculos() {
         return new QuestionarioVeiculos();
@@ -151,7 +175,7 @@ public class ObjectFactory {
 
     /**
      * Create an instance of {@link RespostaAval }
-     * 
+     *
      */
     public RespostaAval createRespostaAval() {
         return new RespostaAval();
@@ -159,7 +183,7 @@ public class ObjectFactory {
 
     /**
      * Create an instance of {@link Veiculo }
-     * 
+     *
      */
     public Veiculo createVeiculo() {
         return new Veiculo();
@@ -167,7 +191,7 @@ public class ObjectFactory {
 
     /**
      * Create an instance of {@link ArrayOfResposta }
-     * 
+     *
      */
     public ArrayOfResposta createArrayOfResposta() {
         return new ArrayOfResposta();
@@ -175,23 +199,15 @@ public class ObjectFactory {
 
     /**
      * Create an instance of {@link Resposta }
-     * 
+     *
      */
     public Resposta createResposta() {
         return new Resposta();
     }
 
     /**
-     * Create an instance of {@link ArrayOfVeiculoQuestao }
-     * 
-     */
-    public ArrayOfVeiculoQuestao createArrayOfVeiculoQuestao() {
-        return new ArrayOfVeiculoQuestao();
-    }
-
-    /**
      * Create an instance of {@link ArrayOfQuestionarioVeiculos }
-     * 
+     *
      */
     public ArrayOfQuestionarioVeiculos createArrayOfQuestionarioVeiculos() {
         return new ArrayOfQuestionarioVeiculos();
@@ -199,15 +215,79 @@ public class ObjectFactory {
 
     /**
      * Create an instance of {@link Questionario }
-     * 
+     *
      */
     public Questionario createQuestionario() {
         return new Questionario();
     }
 
     /**
+     * Create an instance of {@link ArrayOfAvaliacao }
+     *
+     */
+    public ArrayOfAvaliacao createArrayOfAvaliacao() {
+        return new ArrayOfAvaliacao();
+    }
+
+    /**
+     * Create an instance of {@link Avaliacao }
+     *
+     */
+    public Avaliacao createAvaliacao() {
+        return new Avaliacao();
+    }
+
+    /**
+     * Create an instance of {@link ArrayOfVeiculo }
+     *
+     */
+    public ArrayOfVeiculo createArrayOfVeiculo() {
+        return new ArrayOfVeiculo();
+    }
+
+    /**
+     * Create an instance of {@link ArrayOfVeiculoChecklist }
+     *
+     */
+    public ArrayOfVeiculoChecklist createArrayOfVeiculoChecklist() {
+        return new ArrayOfVeiculoChecklist();
+    }
+
+    /**
+     * Create an instance of {@link ArrayOfItemCritico }
+     *
+     */
+    public ArrayOfItemCritico createArrayOfItemCritico() {
+        return new ArrayOfItemCritico();
+    }
+
+    /**
+     * Create an instance of {@link FarolDia }
+     *
+     */
+    public FarolDia createFarolDia() {
+        return new FarolDia();
+    }
+
+    /**
+     * Create an instance of {@link ItemCritico }
+     *
+     */
+    public ItemCritico createItemCritico() {
+        return new ItemCritico();
+    }
+
+    /**
+     * Create an instance of {@link ArrayOfVeiculoQuestao }
+     *
+     */
+    public ArrayOfVeiculoQuestao createArrayOfVeiculoQuestao() {
+        return new ArrayOfVeiculoQuestao();
+    }
+
+    /**
      * Create an instance of {@link ArrayOfRespostaAval }
-     * 
+     *
      */
     public ArrayOfRespostaAval createArrayOfRespostaAval() {
         return new ArrayOfRespostaAval();
@@ -215,15 +295,23 @@ public class ObjectFactory {
 
     /**
      * Create an instance of {@link Questao }
-     * 
+     *
      */
     public Questao createQuestao() {
         return new Questao();
     }
 
     /**
+     * Create an instance of {@link ArrayOfFarolDia }
+     *
+     */
+    public ArrayOfFarolDia createArrayOfFarolDia() {
+        return new ArrayOfFarolDia();
+    }
+
+    /**
      * Create an instance of {@link VeiculoQuestao }
-     * 
+     *
      */
     public VeiculoQuestao createVeiculoQuestao() {
         return new VeiculoQuestao();
@@ -231,18 +319,10 @@ public class ObjectFactory {
 
     /**
      * Create an instance of {@link ArrayOfQuestao }
-     * 
+     *
      */
     public ArrayOfQuestao createArrayOfQuestao() {
         return new ArrayOfQuestao();
-    }
-
-    /**
-     * Create an instance of {@link ArrayOfVeiculo }
-     * 
-     */
-    public ArrayOfVeiculo createArrayOfVeiculo() {
-        return new ArrayOfVeiculo();
     }
 
 }