+++ CHANGELOG.md
@@ -32,6 +32,11 @@ intervalo
 ### Refactor
 * Cria métodos necessários para dar suporte a realização de intervalo offline no aplicativo
 
+### Features
+* Cria métodos para tornar funcional a nova tela de farol do checklist
+* Adiciona atributos no checklist para carregar a quantidade de itens OK/NOK
+* Adiciona métodos para buscar os checklists completos ou resumidos
+
 <a name="v0.0.31"></a>
 ## Version [v0.0.31](https://github.com/luizfp/PrologWebService/compare/v0.0.30...v0.0.31) (2017-09-14)